+++ docs/examples/03_backends_ros/files/02_forward_kinematics_ros_loader.py
@@ -6,10 +6,7 @@ with RosClient() as client:
 
     configuration = Configuration.from_prismatic_and_revolute_values([0.], [-2.238, -1.153, -2.174, 0.185, 0.667, 0.])
 
-    frame_RCF = robot.forward_kinematics(configuration)
-    frame_WCF = robot.to_world_coords(frame_RCF)
+    frame_WCF = robot.forward_kinematics(configuration)
 
-    print("Frame in the robot's coordinate system")
-    print(frame_RCF)
     print("Frame in the world coordinate system")
     print(frame_WCF)