+++ src/compas_fab/backends/ros/planner_backend_moveit.py
@@ -139,15 +139,16 @@ class MoveItPlanner(PlannerBackend):
     # planning services
     # ==========================================================================
 
-    def inverse_kinematics_async(self, callback, errback, frame, base_link, group,
-                                 joint_names, joint_positions, avoid_collisions=True,
+    def inverse_kinematics_async(self, callback, errback, robot, frame, group,
+                                 start_configuration, avoid_collisions=True,
                                  constraints=None, attempts=8, attached_collision_meshes=None):
         """Asynchronous handler of MoveIt IK service."""
+        base_link = robot.model.root.name
         header = Header(frame_id=base_link)
         pose = Pose.from_frame(frame)
         pose_stamped = PoseStamped(header, pose)
         joint_state = JointState(
-            name=joint_names, position=joint_positions, header=header)
+            name=start_configuration.joint_names, position=start_configuration.values, header=header)
         start_state = RobotState(
             joint_state, MultiDOFJointState(header=header))
         if attached_collision_meshes:
@@ -169,13 +170,14 @@ class MoveItPlanner(PlannerBackend):
 
         self.GET_POSITION_IK(self, (ik_request, ), convert_to_positions, errback)
 
-    def forward_kinematics_async(self, callback, errback, joint_positions, base_link,
-                                 group, joint_names, ee_link):
+    def forward_kinematics_async(self, callback, errback, robot, configuration,
+                                 group, ee_link):
         """Asynchronous handler of MoveIt FK service."""
+        base_link = robot.model.root.name
         header = Header(frame_id=base_link)
         fk_link_names = [ee_link]
         joint_state = JointState(
-            name=joint_names, position=joint_positions, header=header)
+            name=configuration.joint_names, position=configuration.values, header=header)
         robot_state = RobotState(
             joint_state, MultiDOFJointState(header=header))
 
@@ -191,14 +193,13 @@ class MoveItPlanner(PlannerBackend):
                                     avoid_collisions, path_constraints,
                                     attached_collision_meshes):
         """Asynchronous handler of MoveIt cartesian motion planner service."""
-        base_link = robot.get_base_link_name(group)
+        base_link = robot.model.root.name  # use world coords
         ee_link = robot.get_end_effector_link_name(group)
 
-        joint_names = robot.get_configurable_joint_names(group)
         header = Header(frame_id=base_link)
         waypoints = [Pose.from_frame(frame) for frame in frames]
         joint_state = JointState(header=header,
-                                 name=joint_names,
+                                 name=start_configuration.joint_names,
                                  position=start_configuration.values)
         start_state = RobotState(joint_state, MultiDOFJointState(header=header))
 
@@ -232,7 +233,7 @@ class MoveItPlanner(PlannerBackend):
 
                 start_state = response.start_state.joint_state
                 start_state_types = robot.get_joint_types_by_names(start_state.name)
-                trajectory.start_configuration = Configuration(start_state.position, start_state_types)
+                trajectory.start_configuration = Configuration(start_state.position, start_state_types, start_state.name)
 
                 callback(trajectory)
 
@@ -254,12 +255,13 @@ class MoveItPlanner(PlannerBackend):
 
         # http://docs.ros.org/jade/api/moveit_core/html/utils_8cpp_source.html
         # TODO: if list of frames (goals) => receive multiple solutions?
-        base_link = robot.get_base_link_name(group)
-        joint_names = robot.get_configurable_joint_names(group)
+        base_link = robot.model.root.name  # use world coords
 
         header = Header(frame_id=base_link)
         joint_state = JointState(
-            header=header, name=joint_names, position=start_configuration.values)
+            header=header,
+            name=start_configuration.joint_names,
+            position=start_configuration.values)
         start_state = RobotState(
             joint_state, MultiDOFJointState(header=header))
         if attached_collision_meshes:
@@ -299,7 +301,7 @@ class MoveItPlanner(PlannerBackend):
 
             start_state = response.trajectory_start.joint_state
             start_state_types = robot.get_joint_types_by_names(start_state.name)
-            trajectory.start_configuration = Configuration(start_state.position, start_state_types)
+            trajectory.start_configuration = Configuration(start_state.position, start_state_types, start_state.name)
 
             callback(trajectory)